refactor: combine survey with course service
